Add files via upload
Creating a database for organizing and storing data received through web application with student data and for fetching the most recent data from database.
The data stored in the DB must be organized and maintained in such a way it cater to the needs of the future. Derive insights from the results to device better strategies and enhance students’ performance from the past data.